Add server validation for adding, editing and deleting, now whole app is type sensitive

+++ src/actions/actions.ts
@@ -1,16 +1,25 @@
 "use server";
 
 import { prisma } from "@/lib/db";
-import { PetType } from "@/lib/types";
+
+import { formSchema, IdSchema } from "@/lib/validations";
 
 import { revalidatePath } from "next/cache";
 
-export async function addPetToDb(pet: Omit<PetType, "id">) {
+export async function addPetToDb(pet: unknown) {
+  const validatedPet = formSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return {
+      message: "Failed to valid data",
+    };
+  }
+
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (err) {
+    console.error(err);
     return {
       message: "Failed to add pet",
     };
@@ -18,27 +27,41 @@ export async function addPetToDb(pet: Omit<PetType, "id">) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePetFromDb(petId: string) {
+export async function deletePetFromDb(petId: unknown) {
+  const validateId = IdSchema.safeParse(petId);
+  if (!validateId) {
+    return {
+      message: "Invalid id type",
+    };
+  }
+
   await prisma.pet.delete({
     where: {
-      id: petId,
+      id: validateId.data,
     },
   });
   revalidatePath("/app", "layout");
 }
 
-export async function editPetInDb(
-  updatedPet: Omit<PetType, "id">,
-  selectedId: string
-) {
+export async function editPetInDb(updatedPet: unknown, selectedId: unknown) {
+  const validatedPet = formSchema.safeParse(updatedPet);
+  const validateId = IdSchema.safeParse(selectedId);
+
+  if (!validateId.success || !validatedPet.success) {
+    return {
+      message: "Failed to valdiate data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: selectedId,
+        id: validateId.data,
       },
-      data: updatedPet,
+      data: validatedPet.data,
     });
-  } catch (error) {
+  } catch (err) {
+    console.error(err);
     return {
       message: "Failed to update pet",
     };
